+++ crabml-core/src/backends/mod.rs
@@ -1,5 +1,2 @@
-pub mod cpu;
 pub mod vulkan;
 pub mod wgpu;
-
-pub use cpu::CpuTensor;
