+++ crabml-core/src/cpu/buf/api.rs
@@ -7,19 +7,19 @@ use super::buf_f16::f16_buf_from_bytes;
 use super::buf_f16::quantize_f32_f16;
 use super::buf_f32::f32_buf_from_bytes;
 use super::buf_f32::vec_dot_f32_f32;
-use crate::backends::cpu::buf::buf_f16::vec_dot_f16_f16;
-use crate::backends::cpu::buf::QuantBufQ2K;
-use crate::backends::cpu::buf::QuantBufQ3K;
-use crate::backends::cpu::buf::QuantBufQ4K;
-use crate::backends::cpu::buf::QuantBufQ4_0;
-use crate::backends::cpu::buf::QuantBufQ4_1;
-use crate::backends::cpu::buf::QuantBufQ5K;
-use crate::backends::cpu::buf::QuantBufQ5_0;
-use crate::backends::cpu::buf::QuantBufQ5_1;
-use crate::backends::cpu::buf::QuantBufQ6K;
-use crate::backends::cpu::buf::QuantBufQ8K;
-use crate::backends::cpu::buf::QuantBufQ8_0;
-use crate::backends::cpu::buf::QuantBufQ8_1;
+use crate::cpu::buf::buf_f16::vec_dot_f16_f16;
+use crate::cpu::buf::QuantBufQ2K;
+use crate::cpu::buf::QuantBufQ3K;
+use crate::cpu::buf::QuantBufQ4K;
+use crate::cpu::buf::QuantBufQ4_0;
+use crate::cpu::buf::QuantBufQ4_1;
+use crate::cpu::buf::QuantBufQ5K;
+use crate::cpu::buf::QuantBufQ5_0;
+use crate::cpu::buf::QuantBufQ5_1;
+use crate::cpu::buf::QuantBufQ6K;
+use crate::cpu::buf::QuantBufQ8K;
+use crate::cpu::buf::QuantBufQ8_0;
+use crate::cpu::buf::QuantBufQ8_1;
 use crate::error::ErrorKind;
 use crate::error::Result;
 use crate::gguf::GGMLType;
