+++ crabml-core/src/cpu/buf/buf_q2_k.rs
@@ -3,7 +3,7 @@ use std::borrow::Cow;
 use half::f16;
 
 use super::QuantBufQ8K;
-use crate::backends::cpu::buf::util::*;
+use crate::cpu::buf::util::*;
 
 /// A q2_k super block of 2-bit quantization
 ///
@@ -135,7 +135,7 @@ impl<'a> QuantBufQ2K<'a> {
     }
 }
 
-use crate::backends::cpu::buf::buf_q8_k::BlockQ8K;
+use crate::cpu::buf::buf_q8_k::BlockQ8K;
 
 pub fn quantize_f32_q2_k(data: &[f32]) -> Vec<BlockQ2K> {
     let mut bs = Vec::with_capacity(data.len() / QK_K);
@@ -253,7 +253,7 @@ pub fn vec_dot_q2_k_q8_k(q2k_bs: &[BlockQ2K], q8k_bs: &[BlockQ8K]) -> f32 {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::backends::cpu::buf::util::tests::*;
+    use crate::cpu::buf::util::tests::*;
 
     const TEST_SIZE: usize = 256;
     const MAX_Q2K_PRODUCT_ERROR: f32 = 0.02;
