+++ crabml-core/src/cpu/buf/buf_q3_k.rs
@@ -3,9 +3,9 @@ use std::ptr;
 
 use half::f16;
 
-use crate::backends::cpu::buf::buf_q8_k::BlockQ8K;
-use crate::backends::cpu::buf::buf_q8_k::QuantBufQ8K;
-use crate::backends::cpu::buf::util::*;
+use crate::cpu::buf::buf_q8_k::BlockQ8K;
+use crate::cpu::buf::buf_q8_k::QuantBufQ8K;
+use crate::cpu::buf::util::*;
 
 /// A q3_k super block of 3-bit quantization
 ///
@@ -324,7 +324,7 @@ pub fn vec_dot_q3_k_q8_k(q3k_bs: &[BlockQ3K], q8k_bs: &[BlockQ8K]) -> f32 {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::backends::cpu::buf::util::tests::*;
+    use crate::cpu::buf::util::tests::*;
 
     const TEST_SIZE: usize = 256;
     const _MAX_Q3K_PRODUCT_ERROR: f32 = 0.02;
