+++ crabml-core/src/cpu/buf/buf_q4_0.rs
@@ -3,7 +3,7 @@ use std::borrow::Cow;
 use half::f16;
 
 use super::QuantBufQ8_0;
-use crate::backends::cpu::buf::buf_q8_0::BlockQ8_0;
+use crate::cpu::buf::buf_q8_0::BlockQ8_0;
 
 #[repr(C, packed)]
 #[derive(Debug, Clone)]
