+++ crabml-core/src/cpu/buf/buf_q5_0.rs
@@ -6,7 +6,7 @@ use byteorder::LittleEndian;
 use half::f16;
 
 use super::QuantBufQ8_0;
-use crate::backends::cpu::buf::buf_q8_0::BlockQ8_0;
+use crate::cpu::buf::buf_q8_0::BlockQ8_0;
 
 #[derive(Debug, Clone)]
 #[repr(C)]
