+++ crabml-core/src/cpu/buf/buf_q8_k.rs
