+++ crabml-core/src/cpu/buf/util.rs
