+++ crabml-core/src/cpu/primitives/matmul_vec.rs
@@ -1,5 +1,5 @@
-use crate::backends::cpu::buf::CpuTensorBuf;
-use crate::backends::cpu::CpuTensorDeviceRef;
+use crate::cpu::buf::CpuTensorBuf;
+use crate::cpu::CpuTensorDeviceRef;
 use crate::tensor::metrics::TimeMetric;
 use crate::tensor::TensorStrider;
