+++ crabml-core/src/cpu/primitives/rope.rs
@@ -1,6 +1,6 @@
 use std::borrow::Cow;
 
-use crate::backends::cpu::buf::CpuTensorBuf;
+use crate::cpu::buf::CpuTensorBuf;
 use crate::error::Result;
 use crate::tensor::RopeMode;
 use crate::tensor::TensorStrider;
