+++ crabml-core/src/cpu/primitives/softmax.rs
@@ -1,6 +1,6 @@
-use crate::backends::cpu::buf::buf_f32::exp_f32_cached;
-use crate::backends::cpu::buf::CpuTensorBuf;
-use crate::backends::cpu::CpuTensorDeviceRef;
+use crate::cpu::buf::buf_f32::exp_f32_cached;
+use crate::cpu::buf::CpuTensorBuf;
+use crate::cpu::CpuTensorDeviceRef;
 use crate::error::ErrorKind;
 use crate::error::Result;
 use crate::gguf::GGMLType;
