+++ crabml-llama2/src/llama2.rs
@@ -528,9 +528,9 @@ impl<'a, T: Tensor> Llama2Runner<T> {
 #[cfg(test)]
 mod tests {
     use approx::assert_relative_eq;
-    use crabml::backends::cpu::CpuTensorDeviceOptions;
     use crabml::backends::wgpu::WgpuTensorDevice;
     use crabml::backends::wgpu::WgpuTensorDeviceOptions;
+    use crabml::cpu::CpuTensorDeviceOptions;
     use crabml::gguf::GGUFFileLoader;
 
     use super::*;
